requirements.txt
runpod>=1.0.0
requests
loguru
streamlit
python-dotenv
